Merge branch lc873-mnist-03-05

Included in this PR:
 - define `eva_t` as `uint64_t` when doing emulation
 - files for `addmm (mm)`, `threshold`, and `sigmoid`. Related tests are skipped for now. Undergrads/Meng in brg is currently working on these kernels.
 - added `HammerBladeGenerator`
 - re-organized HammerBlade host/device code a little bit ... so mostly each kernel lives in its own files.
 - `abs`, `dot`, and `fill_` kernels are implemented. A few helper functions are added to `Copy.h/cpp` to make offloading happen.
 - enable `.hammerblade` for `torch.nn.Module`
 - enable direct printing in python. Now you can pass a HammerBlade tensor to `print()`
 - `strides` in `TensorIterators` are actually important. In certain cases it could be 0, meaning we are adding each element with a single scalar.

+++ c10/hammerblade/emul/bsg_manycore_cuda.h
@@ -294,15 +294,15 @@ typedef int hb_mc_manycore_id_t;
          * @param[in]  argv          List of input arguments to kernel
          * @return HB_MC_SUCCESS if succesful. Otherwise an error code is returned.
          */
-        // Emulation: I changed const uint32_t *argv to uint32_t *argv so it plays
-        // nicely with std::vector
+        // Emulation: I changed const uint32_t *argv to uint64_t *argv so it plays
+        // nicely with std::vector and 64bit memory addr
         __attribute__((warn_unused_result))
         int hb_mc_kernel_enqueue (hb_mc_device_t *device,
                                        hb_mc_dimension_t grid_dim,
                                        hb_mc_dimension_t tg_dim,
                                        const char *name,
                                        const uint32_t argc,
-                                       uint32_t *argv);
+                                       uint64_t *argv);
