+++ c10/hammerblade/emul/kernel_trampoline.cpp
@@ -1,11 +1,11 @@
 #include <kernel_trampoline.h>
 
-std::map<std::string, std::function<int(uint32_t, uint32_t*)>> kernelMap;
-std::vector<std::function<int(uint32_t, uint32_t*)>> enqueued_kernel;
+std::map<std::string, std::function<int(uint32_t, uint64_t*)>> kernelMap;
+std::vector<std::function<int(uint32_t, uint64_t*)>> enqueued_kernel;
 std::vector<uint32_t>  enqueued_argc;
-std::vector<uint32_t*> enqueued_argv;
+std::vector<uint64_t*> enqueued_argv;
 
-void enqueue_kernel(const std::string &kernel, uint32_t argc, uint32_t* argv) {
+void enqueue_kernel(const std::string &kernel, uint32_t argc, uint64_t* argv) {
   assert (kernelMap.find(kernel) != kernelMap.end());
   enqueued_argc.push_back(argc);
   enqueued_argv.push_back(argv);
